+++ changelog.md
@@ -31,6 +31,7 @@ Codion Change Log
 - EntityTableExport now cleans up the file if export is cancelled.
 - EntityTableExportPanel, added save and open configuration.
 - EntityTablePanel.ColumnSelection, AutoResizeModeSelection and EditAttributeSelection enums combined into SelectionMode.
+- EntityTablePanel.Config.includeExport() added along with INCLUDE_EXPORT configuration property, default false.
 
 ## 0.18.51
 ### is.codion.swing.common.ui